PR feedback - moved client4.UploadFiles to _test for now
